# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SEPL Process Scheduler
host: localhost:8080
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /schedule:
    x-swagger-router-controller: schedule
    get:
      description: Returns all existing process schedules to the caller.
      operationId: getJobs
      responses:
        200:
          description: A process schedule array.
          schema:
            type: array
            items:
              $ref: "#/definitions/ScheduleResponse"
        404:
          description: No process schedules were found.
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a new process schedule.
      operationId: addJob
      parameters:
        - in: body
          name: schedule
          required: true
          description: The process schedule to create.
          schema:
            $ref: "#/definitions/ScheduleRequest"
      responses:
        201:
          description: Created
          schema:
              $ref: "#/definitions/ScheduleResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /schedule/{id}:
    x-swagger-router-controller: schedule
    get:
      description: Returns an existing process schedule to the caller.
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The ID of the process schedule to return.
      operationId: getJob
      responses:
        200:
          description: A process schedule object.
          schema:
            $ref: "#/definitions/ScheduleResponse"
        404:
          description: A process schedule with the specified ID was not found.
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates a existing process schedule.
      operationId: updateJob
      parameters:
        - in: body
          name: schedule
          required: true
          description: The process schedule to create.
          schema:
            $ref: "#/definitions/ScheduleRequest"
        - in: path
          name: id
          required: true
          type: string
          description: The ID of the process schedule to update
      responses:
        200:
          description: Updated
          schema:
            type: object
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    delete:
      description: Deletes an existing process schedule.
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The ID of the process schedule to delete.
      operationId: deleteJob
      responses:
        204:
          description: Process schedule deleted successfully.
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  ScheduleResponse:
    properties:
      _id:
        type: string
        example: 522ewb244134affq32
      processDefinitionId:
        type: string
        example: Process:1add31233-124rf5435-fdsfd
      startTime:
        type: string
  ScheduleRequest:
    type: object
    properties:
      processDefinitionId:
        type: string
        example: Process:1add31233-124rf5435-fdsfd
      startTime:
        type: string
    required:
      - processDefinitionId
      - startTime